fix: address 13 round-2 PR review findings for CFO optimizer
- (A) _find_most_used_model accepts pre-filtered agent records
- (B) _find_cheaper_model respects min_context for context window
- (C) recommend_downgrades returns real budget_pressure when no resolver
- (D) evaluate_operation uses projected_alert for conditions
- (E) reports.py logs WARNING before validation ValueErrors
- (F) suggest_routing_optimizations docstring no longer claims latency
- (G) generate_report derives total_cost from records for consistency
- (H) evaluate_operation split into _check_denial/_build_approval_conditions;
      recommend_downgrades/suggest_routing_optimizations loops extracted
- (I) recommend_downgrades parallelizes get_records + budget_pressure
- (J) test_optimizer.py split into 3 files (analysis, decisions)
- (K) DESIGN_SPEC §10.3 mentions routing optimization
- (L) _find_cheaper_model tests exercise actual code path + min_context
