+++ src/ai_company/observability/events/cfo.py
@@ -15,3 +15,4 @@ CFO_REPORT_GENERATOR_CREATED: Final[str] = "cfo.report_generator.created"
 CFO_RESOLVER_MISSING: Final[str] = "cfo.resolver.missing"
 CFO_INSUFFICIENT_WINDOWS: Final[str] = "cfo.anomaly.insufficient_windows"
 CFO_ROUTING_OPTIMIZATION_COMPLETE: Final[str] = "cfo.routing.optimization_complete"
+CFO_REPORT_VALIDATION_ERROR: Final[str] = "cfo.report.validation_error"
