+++ tests/unit/budget/conftest.py
@@ -35,6 +35,8 @@ from ai_company.budget.spending_summary import (
     SpendingSummary,
 )
 from ai_company.budget.tracker import CostTracker
+from ai_company.providers.routing.models import ResolvedModel
+from ai_company.providers.routing.resolver import ModelResolver
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -305,6 +307,66 @@ def make_cost_record(  # noqa: PLR0913
         cost_usd=cost_usd,
         timestamp=timestamp or datetime(2026, 2, 15, 12, 0, 0, tzinfo=UTC),
     )
+
+
+# ── CostOptimizer test helpers ───────────────────────────────────
+
+OPT_START = datetime(2026, 2, 1, tzinfo=UTC)
+OPT_END = datetime(2026, 3, 1, tzinfo=UTC)
+
+
+def make_optimizer(
+    *,
+    budget_config: BudgetConfig | None = None,
+    config: CostOptimizerConfig | None = None,
+    model_resolver: ModelResolver | None = None,
+) -> tuple[CostOptimizer, CostTracker]:
+    """Build a CostOptimizer with a fresh CostTracker."""
+    bc = budget_config or BudgetConfig(total_monthly=100.0)
+    tracker = CostTracker(budget_config=bc)
+    optimizer = CostOptimizer(
+        cost_tracker=tracker,
+        budget_config=bc,
+        config=config,
+        model_resolver=model_resolver,
+    )
+    return optimizer, tracker
+
+
+def make_resolver(
+    models: list[ResolvedModel] | None = None,
+) -> ModelResolver:
+    """Build a ModelResolver from a list of ResolvedModel."""
+    if models is None:
+        models = [
+            ResolvedModel(
+                provider_name="test-provider",
+                model_id="test-large-001",
+                alias="large",
+                cost_per_1k_input=0.03,
+                cost_per_1k_output=0.06,
+            ),
+            ResolvedModel(
+                provider_name="test-provider",
+                model_id="test-medium-001",
+                alias="medium",
+                cost_per_1k_input=0.01,
+                cost_per_1k_output=0.02,
+            ),
+            ResolvedModel(
+                provider_name="test-provider",
+                model_id="test-small-001",
+                alias="small",
+                cost_per_1k_input=0.001,
+                cost_per_1k_output=0.002,
+            ),
+        ]
+    index: dict[str, ResolvedModel] = {}
+    for m in models:
+        index[m.model_id] = m
+        if m.alias is not None:
+            index[m.alias] = m
+    return ModelResolver(index)
 
 
 # ── CFO / CostOptimizer fixtures ─────────────────────────────────
